Demonstrate egregious usage-consuming long-poll loop in Cloud Firestore
As of Firebase SDK 7.9.1, if you delete and then add documents to a
Firestore collection which is being listened to by slow clients, you can
create an unstoppable long-poll loop that fails to reflect the document
removals but continuously counts against your read quotas. It is unclear
whether this is a Firebase SDK or Firestore bug.